requirements.txt
# Core dependencies for Instagram AI Agent
python-dotenv>=1.0.0
requests>=2.31.0
filestack-python>=4.0.0

# Optional AI/ML dependencies (for future enhancement)
openai>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
# Local model providers (optional)
# torch>=2.0.0
# transformers>=4.30.0
# sentence-transformers>=2.2.0

# Development and testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Scheduling and background tasks
schedule>=1.2.0
APScheduler>=3.10.0

# Data processing and analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Logging and monitoring
loguru>=0.7.0
rich>=13.0.0

# Configuration management
pyyaml>=6.0
toml>=0.10.0

# Image processing (for future enhancement)
opencv-python>=4.8.0
Pillow>=10.0.0

# Web framework (for API endpoints - future enhancement)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
